requirement.txt
streamlit
torch==2.2.0+cpu
torchvision==0.17.0+cpu
pillow